+++ app/src/main/java/com/android/mobile_project/data/local/sqlite/dao/UserDao.java
@@ -18,6 +18,6 @@ public interface UserDao {
     void updateUser(UserEntity user);
 
     @Query("SELECT id FROM tbl_user WHERE user_name = :name")
-    public int getUserIdByName(String name);
+    public Long getUserIdByName(String name);
 
 }
